+++ src/Http/Controllers/WebhookController.php
@@ -151,10 +151,14 @@ class WebhookController extends Controller
      */
     protected function handleSubscriptionCreated(array $payload)
     {
-        $passthrough = json_decode($payload['passthrough'], true);
+        $passthrough = isset($payload['passthrough']) ? json_decode($payload['passthrough'], true) : null;
 
-        if (! is_array($passthrough) || ! isset($passthrough['subscription_name'])) {
+        if (! isset($passthrough) || ! is_array($passthrough) || ! isset($passthrough['subscription_name'])) {
             throw new InvalidPassthroughPayload;
+        }
+
+        if ($this->subscriptionExists($payload['subscription_id'])) {
+            return;
         }
 
         $customer = $this->findOrCreateCustomer($payload['passthrough']);
@@ -276,6 +280,17 @@ class WebhookController extends Controller
     protected function findSubscription(string $subscriptionId)
     {
         return Cashier::$subscriptionModel::firstWhere('paddle_id', $subscriptionId);
+    }
+
+    /**
+     * Determine if a subscription with a given Paddle ID already exists.
+     *
+     * @param  string  $subscriptionId
+     * @return bool
+     */
+    protected function subscriptionExists(string $subscriptionId)
+    {
+        return Cashier::$subscriptionModel::where('paddle_id', $subscriptionId)->exists();
     }
 
     /**