+++ src/Receipt.php
@@ -3,10 +3,13 @@
 namespace Laravel\Paddle;
 
 use Illuminate\Database\Eloquent\Model;
+use Laravel\Paddle\Concerns\ManagesAmounts;
 use Money\Currency;
 
 class Receipt extends Model
 {
+    use ManagesAmounts;
+
     /**
      * The attributes that are not mass assignable.
      *
@@ -51,7 +54,7 @@ class Receipt extends Model
      */
     public function amount()
     {
-        return $this->formatAmount((int) ($this->amount * 100));
+        return $this->formatDecimalAmount($this->amount);
     }
 
     /**
@@ -61,7 +64,7 @@ class Receipt extends Model
      */
     public function tax()
     {
-        return $this->formatAmount((int) ($this->tax * 100));
+        return $this->formatDecimalAmount($this->tax);
     }
 
     /**
@@ -72,16 +75,5 @@ class Receipt extends Model
     public function currency(): Currency
     {
         return new Currency($this->currency);
-    }
-
-    /**
-     * Format the given amount into a displayable currency.
-     *
-     * @param  int  $amount
-     * @return string
-     */
-    protected function formatAmount($amount)
-    {
-        return Cashier::formatAmount($amount, $this->currency);
     }
 }