+++ src/Subscription.php
@@ -445,6 +445,10 @@ class Subscription extends Model
     {
         $this->guardAgainstUpdates('update quantities');
 
+        if ($quantity < 1) {
+            throw new LogicException('Paddle does not allow subscriptions to have a quantity of zero.');
+        }
+
         $this->updatePaddleSubscription(array_merge($options, [
             'quantity' => $quantity,
             'prorate' => $this->prorate,
@@ -565,7 +569,7 @@ class Subscription extends Model
     /**
      * Get the Paddle update url.
      *
-     * @return array
+     * @return string
      */
     public function updateUrl()
     {
@@ -673,7 +677,7 @@ class Subscription extends Model
     /**
      * Get the Paddle cancellation url.
      *
-     * @return array
+     * @return string
      */
     public function cancelUrl()
     {