+++ tests/Feature/PaymentTest.php
@@ -17,6 +17,24 @@ class PaymentTest extends FeatureTestCase
         $this->assertSame('EUR', $payment->currency()->getCode());
     }
 
+    public function test_it_can_returns_a_japanese_amount_and_currency()
+    {
+        $payment = new Payment('1200.0', 'JPY', '2020-05-07');
+
+        $this->assertSame('¥1,200', $payment->amount());
+        $this->assertSame('1200.0', $payment->rawAmount());
+        $this->assertSame('JPY', $payment->currency()->getCode());
+    }
+
+    public function test_it_can_return_a_korean_amount_and_currency()
+    {
+        $payment = new Payment('1200.0', 'KRW', '2020-05-07');
+
+        $this->assertSame('₩1,200', $payment->amount());
+        $this->assertSame('1200.0', $payment->rawAmount());
+        $this->assertSame('KRW', $payment->currency()->getCode());
+    }
+
     public function test_it_can_be_serialized_to_an_array()
     {
         $payment = new Payment('12.45', 'EUR', '2020-05-07');